fix(llm): fix llm-create, add llm-list & llm-show

+++ pkg/compute/models/containers.go
@@ -781,6 +781,11 @@ func (c *SContainer) StartPullImageTask(ctx context.Context, userCred mcclient.T
 }
 
 func (c *SContainer) RealDelete(ctx context.Context, userCred mcclient.TokenCredential) error {
+	// delete related llm
+	if err := LLMManager.DeleteByContainerId(ctx, userCred, c.GetId()); nil != err {
+		return err
+	}
+
 	return c.SVirtualResourceBase.Delete(ctx, userCred)
 }
